test: split App tests to components
